chore: move @shared folder into the src
