+++ src/@shared/src/domain/contracts/entity/entity-base.ts
