+++ src/@shared/src/domain/contracts/infra/repository/repository-base.ts
@@ -1,5 +1,5 @@
-import { EntityBase } from '@shared/src/domain/models/entities/entity-base';
-import { ValueObject } from '@shared/src/domain/value-objects/value-object';
+import { EntityBase } from 'src/@shared/src/domain/models/entities/entity-base';
+import { ValueObject } from 'src/@shared/src/domain/value-objects/value-object';
 import IEntityBase from '../../entity/entity-base';
 import { SearchParams } from './search-params';
 import { SearchResult } from './search-result';
