+++ src/@shared/src/domain/errors/not-found.error.ts
