+++ src/@shared/src/domain/models/entities/entity-base.ts
