+++ src/@shared/src/domain/use-cases/use-case.ts
