+++ src/@shared/src/domain/value-objects/uuid.vo.ts
