+++ src/@shared/src/domain/value-objects/value-object.ts
