+++ src/@shared/src/infra/config.ts
