+++ src/core/category/update/update-category.use-case.ts
@@ -1,4 +1,4 @@
-import { NotFoundError } from '@shared/src/domain/errors/not-found.error';
+import { NotFoundError } from 'src/@shared/src/domain/errors/not-found.error';
 
 import { ICategoryRepository } from 'src/core/domain/contracts/repositories/category.repository';
 import {
