+++ src/core/domain/contracts/use-cases/category/create/create-category.ts
@@ -1,4 +1,4 @@
-import { IUseCase } from '@shared/src/domain/use-cases/use-case';
+import { IUseCase } from 'src/@shared/src/domain/use-cases/use-case';
 import { CategoryOutputMapper } from 'src/core/category/common/category-output';
 
 export type ICreateCategoryInput = {
