+++ src/core/domain/contracts/use-cases/category/delete/delete-category.use-case.ts
@@ -1,4 +1,4 @@
-import { IUseCase } from '@shared/src/domain/use-cases/use-case';
+import { IUseCase } from 'src/@shared/src/domain/use-cases/use-case';
 
 export type IDeleteCategoryInput = {
   id: string;
