+++ src/core/domain/contracts/use-cases/category/getAll/getAll-category.ts
@@ -1,4 +1,4 @@
-import { IUseCase } from '@shared/src/domain/use-cases/use-case';
+import { IUseCase } from 'src/@shared/src/domain/use-cases/use-case';
 import { Category } from 'src/core/domain/entities/category.entity';
 
 export abstract class IGetAllCategoryUseCase
