+++ src/core/domain/entities/category.entity.ts
@@ -1,5 +1,5 @@
-import { Uuid } from '@shared/src/domain/value-objects/uuid.vo';
-import { ValueObject } from '@shared/src/domain/value-objects/value-object';
+import { Uuid } from 'src/@shared/src/domain/value-objects/uuid.vo';
+import { ValueObject } from 'src/@shared/src/domain/value-objects/value-object';
 
 export type CategoryConstructorProps = {
   category_id?: CategoryId;
