+++ test/data/use-cases/category/delete-category.use-case.spec.ts
@@ -1,5 +1,5 @@
-import { NotFoundError } from '@shared/src/domain/errors/not-found.error';
-import { setupSequelize } from '@shared/src/infra/test/helpers';
+import { NotFoundError } from 'src/@shared/src/domain/errors/not-found.error';
+import { setupSequelize } from 'src/@shared/src/infra/test/helpers';
 import { DeleteCategoryUseCase } from 'src/core/category/delete/delete-category.use-case';
 import { Category, CategoryId } from 'src/core/domain/entities/category.entity';
 import { CategoryModel } from 'src/core/infra/db/postgres/category/category.model';
